feat: change a lot

+++ docker/restore.py
@@ -1,20 +1,23 @@
 import os
+import re
 import sys
 from datetime import datetime, timedelta
 import subprocess
 
 # 数据库还原
 
-BACKUP_PATH = "/backup"
+DEFAULT_BACKUP_PATH = "/backup"
+DEFAULT_BACKUP_SAVE_NUMS = 3
 
-must_inputs = ["MONGO_URI", "MONGO_FILE_PREFIX", "BACKUP_LATEST_FILE"]
+must_inputs = ["MONGO_URI"]
 other_inputs = [
+    "FILE_PREFIX",
     "BACKUP_PATH",
-    "BACKUP_PWD"
+    "BACKUP_PWD",
 ]
 
 for key in must_inputs:
-    if key not in os.environ or os.environ[key] == '':
+    if key not in os.environ or os.environ[key] == "":
         print(key, "must set")
         sys.exit()
 
@@ -27,8 +30,6 @@ def check_var(key):
 
 # 必填
 uri = os.environ["MONGO_URI"]
-file_prefix = os.environ["MONGO_FILE_PREFIX"]
-backup_latest_file = os.environ["BACKUP_LATEST_FILE"]
 
 # 选填
 backup_path = os.environ["BACKUP_PATH"] if check_var("BACKUP_PATH") else BACKUP_PATH
@@ -38,55 +39,55 @@ date = (datetime.utcnow() + timedelta(hours=8)).strftime("%Y%m%d%H%M%S")
 
 
 def get_files():
-    file = os.listdir(backup_path)
-    temp_files = []
-    latest_file_name = f'{backup_latest_file}.zip' if backup_pwd else backup_latest_file
-    for file_name in file:
-        file_path = os.path.join(backup_path, file_name)
-        if os.path.isfile(file_path):
-            # 有密码备份文件为tar.gz.zip格式，无密码备份文件为tar.gz格式
-            if (backup_pwd and file_name.endswith('.zip')) or (not backup_pwd and file_name.endswith('.tar.gz')):
-                if file_name.startswith(file_prefix) and file_name != latest_file_name:
-                    temp_files.append({
-                        "time": int(file_name.split('.')[0].split('-')[-1]),
-                        "path": file_path,
-                        "name": file_name
-                    })
-    files = sorted(temp_files, key=lambda d: d['time'], reverse=True)
-    files.insert(0, {
-        "path": os.path.join(backup_path, latest_file_name),
-        "name": latest_file_name
-    })
-    return files
+    # 构造正则表达式
+    regex_pattern = f"^.*(\\d{{14}})\\.tar\\.gz(\\.crypt)?$"
+    compiled_regex = re.compile(regex_pattern)
+
+    # 列出备份文件夹中所有的文件
+    all_files = os.listdir(backup_path)
+
+    # 筛选出符合正则表达式的文件
+    matched_files = [file for file in all_files if compiled_regex.match(file)]
+
+    # 根据文件名排序（这使得最新的备份文件位于列表的末尾）
+    matched_files.sort(reverse=True)
+    return matched_files
 
 
 def restore_file(file_path):
-    restore_path = ".".join(file_path.split('.')[:-1]) if backup_pwd else file_path
-    if backup_pwd:
+    is_crypt_file = file_path.endswith(".crypt")
+    restore_path = file_path if not is_crypt_file else file_path[:-6]
+
+    if is_crypt_file and not backup_pwd:
+        print("need set BACKUP_PWD")
+        sys.exit()
+
+    if is_crypt_file and backup_pwd:
         # 先解密
         unzip_path = os.path.dirname(file_path)
         # -o 覆盖已有文件，-j 不保留文件夹
-        subprocess.call(f'unzip -P {backup_pwd} -oj {file_path} -d {unzip_path}', shell=True)
+        subprocess.call(f"unzip -P {backup_pwd} -oj {file_path} -d {unzip_path}", shell=True)
 
     # 恢复数据
-    cmd = f'mongorestore --uri="{uri}" --gzip --archive={restore_path} --authenticationDatabase=admin'
+    cmd = f'mongorestore --uri="{uri}" --gzip --archive={restore_path}'
     subprocess.call(cmd, shell=True)
 
-    if backup_pwd:
+    if is_crypt_file and backup_pwd:
         # 删除解密文件
         os.remove(restore_path)
 
 
 try:
-    # 1. 清理备份，保留最近的若干份
+    # 1. 获取备份文件列表
     backup_files = get_files()
 
     input_cmd = "请选择要还原的备份文件：\n"
     for index, item in enumerate(backup_files):
-        input_cmd += f'{index + 1}. {item["name"]}\n'
+        input_cmd += f"{index + 1}. {item}\n"
     value = input(input_cmd)
+    file = f"{backup_path}/{backup_files[int(value) - 1]}"
 
-    restore_file(backup_files[int(value) - 1]["path"])
+    restore_file(file)
 
     print("script end")
 except Exception as e:
